requirements.txt
asgiref==3.7.2
boto3==1.34.38
botocore==1.34.43
django-cors-headers==4.3.1
django-jazzmin==2.6.0
django-nested-admin==4.0.2
django-nested-inline==0.4.6
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
jmespath==1.0.1
pillow==10.2.0
psycopg2-binary
PyJWT==2.8.0
python-dateutil==2.8.2
python-monkey-business==1.0.0
pytz==2024.1
s3transfer==0.10.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2024.1
urllib3
whitenoise==6.6.0
